// libs/modules/sps-website-builder/models/navbar/frontend/component/variants/sps-lite/admin-form/src/lib/interface.ts
import { IModel } from "@sps/sps-website-builder-models-navbar-contracts";
import { IModel as IModelExtended } from "@sps/sps-website-builder-models-navbar-contracts-extended";

export const variant = "admin-form" as const;

export interface IComponentBase {
  showSkeletons?: boolean;
  isServer: boolean;
}

export interface IComponentProps extends IComponentBase {
  variant: typeof variant;
  className?: string;
  data?: IModel;
  setOpen?: (open: boolean) => void;
}

export interface IComponentPropsExtended extends IComponentProps {
  data?: IModelExtended;
}